feat: add API to reload files of a namespace
To supplement Drag'n'Drop functionality for file-upload (which is
implemented by the dropfiles plugin), it is necessary that we can update
the files shown in a namespace.

This commit adds a custom event to the filelisting table, which reloads
the data of a given namespace. This event can than be triggered by
external javascript, like the javascript of the dropfiles plugin.