form for deleting files
The commit adds checkboxes for deleting files. The delete button appears at the left of filter input. If no files are selected, the button is hidden. The checkboxes column is always visible even if a user doesn't have a DELETE permission on a root namespace, since he can have a permission on some subnamespaces.